backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymysql
python-keycloak
python-jose[cryptography]